# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Profile:
    properties:
      age:
        type: integer
      avatar:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      friends:
        type: integer
      gender:
        type: string
      id:
        type: integer
      level:
        type: integer
      losses:
        type: integer
      medals:
        items:
          type: string
        type: array
      name:
        type: string
      play_time:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      wins:
        type: integer
    type: object
  domain.PurchaseItemRequest:
    properties:
      item_id:
        type: integer
    required:
    - item_id
    type: object
  domain.PurchaseRequest:
    properties:
      plan_id:
        type: string
    type: object
...
  domain.ShopItem:
    properties:
      created_at:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
  domain.UpdateProfileRequest:
    properties:
      avatar:
        type: string
      name:
        type: string
    type: object
...
  domain.Wallet:
    properties:
      coins:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      diamonds:
        type: integer
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Mafia game.
  title: Mafia API
  version: "1.0"
paths:
...
  /admin/shop/items:
    post:
      consumes:
      - application/json
      description: Adds a new item to the shop catalog.
      parameters:
      - description: Shop item payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ShopItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ShopItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a shop item
      tags:
      - Admin
  /admin/shop/items/{id}:
    delete:
      description: Deletes an existing shop item by ID.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a shop item
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Updates an existing shop item by ID.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Shop item payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ShopItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ShopItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a shop item
      tags:
      - Admin
...
  /shop/items:
    get:
      description: Returns all shop items available for purchase.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ShopItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List shop items
      tags:
      - Shop
  /shop/purchase:
    post:
      consumes:
      - application/json
      description: Purchases an item from the shop for the authenticated user.
      parameters:
      - description: Purchase payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.PurchaseItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ShopItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Purchase a shop item
      tags:
      - Shop
  /user/dashboard:
    get:
      description: Retrieves aggregated dashboard information for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user dashboard
      tags:
      - User
  /user/profile:
    get:
      description: Retrieves the authenticated user's profile details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Profile'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's profile information.
      parameters:
      - description: Profile update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Profile'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
  /user/purchase:
    post:
      consumes:
      - application/json
      description: Initiates a purchase flow for the authenticated user.
      parameters:
      - description: Purchase payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.PurchaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Initiate wallet purchase
      tags:
      - User
  /user/wallet:
    get:
      description: Retrieves the authenticated user's wallet balances.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Wallet'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get wallet details
      tags:
      - User
swagger: "2.0"